libgloss/kvx-cos: fix stat syscall
Summary:
To be aligned with changed introduced by D2867 for the fstat syscall,
use a uint64_t buffer for struct stat content passing for the stat
syscall as well.


Test Plan: CI

Reviewers: O26 newlib!

Differential Revision: https://phab.kalray.eu/D13939

+++ libgloss/kvx-cos/stat.c
@@ -13,7 +13,7 @@ int sc_stat(const char *filename, int *);
 int _stat(const char *filename, struct stat *st)
 {
     int ret;
-    int res[13];
+    uint64_t res[13];
 
     ret = sc_stat(filename, res);
